% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_arrow_plotting_df.R
\name{create_arrow_plotting_df}
\alias{create_arrow_plotting_df}
\title{Create df for plotting arrows.}
\usage{
create_arrow_plotting_df(df, middle, width, plasmid_length, arrowhead_size = 8)
}
\description{
Create df for plotting arrows.
}
