% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_feature_df.R
\name{create_feature_df}
\alias{create_feature_df}
\title{Extract features into a dataframe.}
\usage{
create_feature_df(input)
}
\arguments{
\item{input}{\code{readlines()} of a \code{.gb} file.}
}
\description{
Parses lines of text from a .gb format plasmid sequence file. Detects
features and details for the features from the file and returns it in a data
frame to for use in creating arrows.
}
