% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_sequence.R
\name{extract_sequence}
\alias{extract_sequence}
\title{Extract sequence from readlines(plasmid)}
\usage{
extract_sequence(plasmid)
}
\arguments{
\item{plasmid}{Takes result of \code{readLines()} and parses it to extract
the final plasmid sequence. Currently only used to final the plasmid
length.}
}
\description{
Extract sequence from readlines(plasmid)
}
