% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_features.R
\name{get_features}
\alias{get_features}
\title{Extract Lines that Contain Feature Information}
\usage{
get_features(plasmid)
}
\arguments{
\item{plasmid}{Result of \code{readLines()}. Parses the file to get features
from the read .gb file.}
}
\description{
Extract Lines that Contain Feature Information
}
