% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_fit_text.R
\name{makeContent.fittexttree}
\alias{makeContent.fittexttree}
\title{Used Internally}
\usage{
\method{makeContent}{fittexttree}(x)
}
\arguments{
\item{x}{Fittexttree, used internally.}
}
\description{
Used Internally
}
