% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_plasmid.R
\name{parse_plasmid}
\alias{parse_plasmid}
\title{Parse Plasmid}
\usage{
parse_plasmid(file)
}
\arguments{
\item{file}{A file containing a \code{.gb} Genbank formatted plasmid
sequence. File should contain list of features in standard Genbank
notation.}
}
\description{
Parses a \code{.gb} Genbank file. Returns a list containing a dataframe of
features, the sequence, and the length of the plasmid for use in the
\code{render_plasmid()} function.
}
