% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plasmid_gadget.R
\name{plasmid_gadget}
\alias{plasmid_gadget}
\title{Gadget Function}
\usage{
plasmid_gadget(plasmid)
}
\arguments{
\item{plasmid}{Object created by the \code{parse_plasmid()} function.}
}
\description{
Creates interactive shiny gadget to easily adjust easthetic parameters of the
final plot. The app will return the code required to re-create the currently
viewed plot.
}
