% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plasmid_plot.R
\name{plasmid_plot}
\alias{plasmid_plot}
\title{Plasmid Plotting Function}
\usage{
plasmid_plot(
  arrow_df,
  labels,
  angle_adjustment,
  zoom_y,
  plasmidName,
  font_family = "mono",
  nameSize,
  curve,
  labelNudge,
  labelSize,
  repelBox,
  plasmid_length,
  plasmid_y = 2,
  plasmid_width = 0.1
)
}
\arguments{
\item{arrow_df}{Dataframe of points from the
\code{create_arrow_plotting_df()} function to draw arrows.}

\item{labels}{List resulting from \code{create_labels()} function containing
curved text dataframe and labels for annotation.}

\item{angle_adjustment}{Angle (in degrees) to rotate the resulting plasmid
map by.}

\item{zoom_y}{Zoom level for the resulting plasmid map. Sets
\code{ylim(0,zoom_y)} and can be overriden.}

\item{plasmidName}{String name for the plasmid map.}

\item{font_family}{Font family to use in all of the text on the plot.}

\item{nameSize}{Size of the plasmid name annotation.}

\item{curve}{Curvature of the repelled text labels.}

\item{labelNudge}{Amount to nudge the labels away from the plasmid arrows to
allow for neater packing.}

\item{labelSize}{Size of the repelled labels.}

\item{repelBox}{Distance to repel between the labels.}

\item{plasmid_length}{Total length of the plasmid.}

\item{plasmid_y}{Position on the y axis of the original plasmid.}

\item{plasmid_width}{Width of the arrows of the plasmid.}
}
\description{
Plotting function to create the final plasmid map from properly file.
}
