% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_plasmid}
\alias{plot_plasmid}
\title{Plot a Plasmid}
\usage{
plot_plasmid(plasmid, name = "Plasmid Name", label_wrap = 20)
}
\arguments{
\item{plasmid}{A list of class 'plasmid' created through \code{read_gb()}.}

\item{name}{Name of the plasmid, to be shown in the center.}

\item{label_wrap}{Passed to \code{stringr::str_wrap()} to wrap the long labels.}
}
\value{
A ggplot object.
}
\description{
Create a \code{{ggplot2}} plot of a plasmid in polar coordinates. Extracts the
features as a data.frame from the plasmid and uses these to construct arrows
that are added to the plot.
}
