% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_text.R
\name{split_text_df}
\alias{split_text_df}
\title{Create dataframe containing split and spaced text.}
\usage{
split_text_df(feat_length, feat_middle, name, char_spacing, reverse = FALSE)
}
\description{
Create dataframe containing split and spaced text.
}
