% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_arrow.R
\name{stat_arrow}
\alias{stat_arrow}
\title{Custom Stat: Arrow}
\usage{
stat_arrow(
  mapping = NULL,
  data = NULL,
  geom = "polygon",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...,
  bp = 6000,
  arrowhead_size = 8
)
}
\arguments{
\item{mapping}{The aesthetic mapping, usually constructed with aes().}

\item{data}{The dataset to be used in the layer.}

\item{geom}{The geometric object to use for the layer, default is "polygon".}

\item{position}{Position adjustment, default is "identity".}

\item{na.rm}{Should missing values be removed?}

\item{show.legend}{Logical. Should this layer be included in the legends?}

\item{inherit.aes}{Should inherit aesthetics from the parent plot?}

\item{...}{Other arguments passed to layer().}

\item{bp}{The base parameter for the arrow shape.}

\item{arrowhead_size}{The size of the arrowhead.}
}
\value{
A ggplot layer with arrow shapes.
}
\description{
This stat creates arrow shapes in a ggplot.
}
\details{
Used for drawing features for a plasmid that work in both cartesian and
polad coordinate systems.
}
\note{
The \code{bp} parameter determines the base parameter of the arrow shape.
The \code{arrowhead_size} parameter controls the size of the arrowhead.
}
\seealso{
\code{\link[ggplot2]{geom_segment}}
}
\keyword{plotting}
