% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_arrow.R
\docType{data}
\name{StatArrow}
\alias{StatArrow}
\title{Compute positions for an Arrow}
\format{
An object of class \code{StatArrow} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
StatArrow
}
\description{
Compute positions for an Arrow
}
\keyword{datasets}
