% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_arrow.R
\docType{data}
\name{StatArrowLabel}
\alias{StatArrowLabel}
\title{Compute required values for labels}
\format{
An object of class \code{StatArrowLabel} (inherits from \code{StatArrow}, \code{Stat}, \code{ggproto}, \code{gg}) of length 2.
}
\usage{
StatArrowLabel
}
\description{
Compute required values for labels
}
\keyword{datasets}
