% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{as.data.frame.plasmid}
\alias{as.data.frame.plasmid}
\title{Extract Features of a Plasmid as a DataFrame}
\usage{
\method{as.data.frame}{plasmid}(x, row.names, optional, ...)
}
\arguments{
\item{x}{A list of class 'plasmid' from \code{read_gb()}}

\item{row.names}{Ignored.}

\item{optional}{Ignored.}

\item{...}{Ignored.}
}
\value{
a DataFrame
}
\description{
Extract Features of a Plasmid as a DataFrame
}
