% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_arrow_positions.R
\name{create_arrow_positions}
\alias{create_arrow_positions}
\title{Create single arrow points df}
\usage{
create_arrow_positions(
  start,
  end,
  middle,
  width,
  plasmid_length,
  arrowhead_size = 8
)
}
\description{
Create single arrow points df
}
