% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plasmid_plot.R
\name{plasmid_plot}
\alias{plasmid_plot}
\title{Plasmid Plotting Function}
\usage{
plasmid_plot(
  arrow_df,
  labels,
  angle_adjustment,
  zoom_y,
  plasmid_name,
  font_family = "mono",
  name_size,
  bp_count = TRUE,
  curved_size,
  curve,
  label_nudge,
  label_size,
  repel_box,
  plasmid_length,
  plasmid_y = 2,
  plasmid_width = 0.1
)
}
\arguments{
\item{arrow_df}{Dataframe of points from the
\code{create_arrow_plotting_df()} function to draw arrows.}

\item{labels}{List resulting from \code{create_labels()} function containing
curved text dataframe and labels for annotation.}

\item{angle_adjustment}{Angle (in degrees) to rotate the resulting plasmid
map by.}

\item{zoom_y}{Zoom level for the resulting plasmid map. Sets
\code{ylim(0,zoom_y)} and can be overriden.}

\item{plasmid_name}{String name for the plasmid map.}

\item{font_family}{Font family to use in all of the text on the plot.}

\item{name_size}{Size of the plasmid name annotation.}

\item{curved_size}{Font size of the curved labels (in .pt).}

\item{curve}{Curvature of the repelled text labels.}

\item{label_nudge}{Amount to nudge the labels away from the plasmid arrows to
allow for neater packing.}

\item{label_size}{Size of the repelled labels.}

\item{repel_box}{Distance to repel between the labels.}

\item{plasmid_length}{Total length of the plasmid.}

\item{plasmid_y}{Position on the y axis of the original plasmid.}

\item{plasmid_width}{Width of the arrows of the plasmid.}
}
\description{
Plotting function to create the final plasmid map from properly file.
}
