% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plasmid_plot.R
\name{plasmid_plot}
\alias{plasmid_plot}
\title{Plasmid Plotting Function}
\usage{
plasmid_plot(
  feature_df,
  features = NULL,
  zoom_y,
  plasmidName,
  nameSize,
  curve,
  rotation,
  labelNudge,
  labelSize,
  repelBox,
  plasmid_length,
  plasmid_y = 2,
  plasmid_width = 0.1,
  arrowhead_size = 4,
  spacing_scale = 0.006,
  label_hjust = 0.4,
  label_length_cutoff = 0.85
)
}
\description{
Plasmid Plotting Function
}
