% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_plasmap.R
\name{render_plasmap}
\alias{render_plasmap}
\title{Render Plasmap}
\usage{
render_plasmap(
  plasmid,
  middle = 2,
  width = 0.1,
  arrowhead_size = 4,
  rotation = 0,
  spacing_scale = 0.006,
  label_hjust = 0.4,
  label_length_cutoff = 0.85,
  label_curve = 0,
  zoom_y = 3,
  plasmid_name = "plasmid_name",
  nameSize = 6,
  curve = 10,
  labelNudge = 0.8,
  labelSize = 10,
  repelBox = 0.2
)
}
\arguments{
\item{plasmid}{A list containing at least a dataframe of features and the
plasmid length.}
}
\description{
Render Plasmap
}
