% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_plasmap.R
\name{render_plasmap}
\alias{render_plasmap}
\title{Render Plasmap}
\usage{
render_plasmap(
  plasmid,
  middle = 2,
  width = 0.1,
  arrowhead_size = 4,
  rotation = 0,
  curved_scaling = 1,
  label_hjust = 0.4,
  label_length_cutoff = 0.85,
  label_curve = 0,
  zoom_y = 3,
  plasmid_name = "plasmid_name",
  name_size = 6,
  curve = 10,
  label_nudge = 0.8,
  repel_box = 0.2,
  bp_count = TRUE
)
}
\arguments{
\item{plasmid}{A list containing at least a dataframe of features and the
plasmid length.}

\item{arrowhead_size}{Size (in degrees) of the arrowheads. Sets how much of
the circle they will take up (total of circle being 360 degrees).}

\item{label_hjust}{Adjusts justification of the curved labels towards the
start (0) or end (1) of the arrow.}

\item{label_length_cutoff}{Proportion of the arrow the text is allowed to
take up before becoming a rep}

\item{spacing_scale}{Scalar for the spacing between the curved text on the
arrows.}
}
\description{
Render Plasmap
}
